Clean up

+++ dust-lang/src/vm/call_frame.rs
@@ -5,6 +5,5 @@ pub struct CallFrame {
     pub register_range_end: u16,
     pub argument_count: u16,
     pub return_count: u16,
-
-    pub instruction_pointer: usize,
+    pub instruction_pointer: u32,
 }
